unsigned problems
git-svn-id: https://svn.uvt.nl/ilk/trunk/sources/Tadpole@2561 12f355fe-0486-481a-ad91-c297ab22b4e3

+++ src/mbma_mod.cxx
@@ -48,8 +48,8 @@ using namespace Timbl;
 int mbaDebug = 0; // debugging on (1) or off (0)
 
 namespace Mbma {
-  const unsigned int LEFT =  6; // left context
-  const unsigned int RIGHT = 6; // right context
+  const long int LEFT =  6; // left context
+  const long int RIGHT = 6; // right context
   const unsigned int NRTAGS = 13; // number of CELEX POS tags
   const int NRINFS = 21; // number of inflection types
